+++ libft/src/ydict/ydict_property.c
@@ -6,7 +6,7 @@
 /*   By: youkim < youkim@student.42seoul.kr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/23 15:46:37 by youkim            #+#    #+#             */
-/*   Updated: 2021/11/23 18:40:33 by youkim           ###   ########.fr       */
+/*   Updated: 2021/11/26 11:48:53 by youkim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 //	visualize list. only works when EVERY values are string.
 void	ydict_list_items(t_dict *dict)
 {
-	size_t		id;
+	int			id;
 	t_dictitem	*item;
 
 	id = 0;
@@ -24,9 +24,9 @@ void	ydict_list_items(t_dict *dict)
 	{
 		item = dict->items[id];
 		if (item && item->key)
-			printf("%s%2zu |%-8s|%s\n%s",
+			printf("%s%2d |%-8s|%s\n%s",
 				HGRN, id++, item->key, item->value, END);
 		else
-			printf("%2zu |%8s|\n", id++, "");
+			printf("%2d |%8s|\n", id++, "");
 	}
 }
