+++ libft/src/ydict/ydict_status.c
@@ -6,7 +6,7 @@
 /*   By: youkim < youkim@student.42seoul.kr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/23 11:17:16 by youkim            #+#    #+#             */
-/*   Updated: 2021/11/23 17:10:17 by youkim           ###   ########.fr       */
+/*   Updated: 2021/11/26 11:28:41 by youkim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,12 +17,12 @@ bool	is_input_valid(t_dict *dict, char *key, void *value)
 	return (dict && key && value);
 }
 
-bool	is_key_vacant(t_dict *dict, size_t id)
+bool	is_key_vacant(t_dict *dict, int id)
 {
 	return (!dict->items[id]);
 }
 
-bool	is_key_update(t_dict *dict, size_t id, char *key)
+bool	is_key_update(t_dict *dict, int id, char *key)
 {
 	return (ystrequ(dict->items[id]->key, key));
 }
@@ -34,7 +34,7 @@ bool	is_dict_almostfull(t_dict *dict)
 
 bool	is_capacity_overflow(t_dict *dict)
 {
-	size_t	new_capacity;
+	int	new_capacity;
 
 	new_capacity = dict->capacity * 2;
 	return (new_capacity < dict->capacity);
